ugt-terminal/worker-broker/pyproject.toml
[project]
name = "worker-broker"
version = "0.1.0"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "binance-connector>=3.5.0",
    "coinbase-advanced-py>=1.2.0",
    "ib_insync>=0.9.86",
    "robin_stocks>=3.1.0",
    "cryptography>=41.0.0",
    "aiohttp>=3.8.0",
    "redis>=5.0.0",
    "asyncpg>=0.28.0",
    "apscheduler>=3.10.0",
    "structlog>=23.1.0",
    "pydantic-settings>=2.0.0"
]

[project.optional-dependencies]
all = [
    "pytest>=7.3.0",
    "pytest-asyncio>=0.21.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
